backend/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.31
python-dotenv==0.21.0
Werkzeug==2.2.2